created a centralized API file
